LP64
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@962 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCMPix.m
@@ -69,9 +69,9 @@ extern void vmultiplyNoAltivec( float *a,  float *b,  float *r, long size);
 
 void ConvertFloatToNative (float *theFloat)
 {
-	unsigned long		*myLongPtr;
+	unsigned int		*myLongPtr;
 	
-	myLongPtr = (unsigned long *)theFloat;
+	myLongPtr = (unsigned int *)theFloat;
 	*myLongPtr = EndianU32_LtoN(*myLongPtr);
 }
 
@@ -6568,11 +6568,11 @@ BOOL            readable = YES;
 							
 							case 8:
 							{
-								unsigned long   *bufPtr;
+								unsigned int   *bufPtr;
 								short			*ptr, *tmpImage;
 								long			loop;
 								
-								bufPtr = (unsigned long*) [fileData bytes]+ frameNo * (realheight * realwidth)*4;
+								bufPtr = (unsigned int*) [fileData bytes]+ frameNo * (realheight * realwidth)*4;
 								ptr    = oImage;
 								
 								loop = totSize/2;
